Extract config validators from prod_validator
